TP-502: And a frontend for the importer and make the templates more generic
It is required that a user can login to the tool and upload a TARIC XML
file to be loaded. This commit introduces a frontend at `/importer` which
allows this functionality.

To facilitate these views the templates for additional codes have been made
more generic so that they can be used across the app. Likewise the filtering
capabilities have been updated to allow handling of GDS crispy forms formatting.
