+++ importer/jinja2/importer/list.jinja
@@ -1,0 +1,7 @@
+{% set object_type = "import batch" %}
+{% set object_type_plural = "import batches" %}
+{% set form_url = "import_batch-ui-list" %}
+{% set list_include = "includes/importer_list.jinja" %}
+
+
+{%- include "layouts/list.jinja" -%}
