+++ importer/views.py
@@ -1,0 +1,48 @@
+from django.db.models import Count
+from django.db.models import Q
+from django.urls import reverse_lazy
+from django.views.generic import FormView
+
+from common.views import RequiresSuperuserMixin
+from common.views import WithPaginationListView
+from importer import forms
+from importer import models
+from importer.filters import ImportBatchFilter
+
+
+class ImportBatchList(RequiresSuperuserMixin, WithPaginationListView):
+    """
+    UI endpoint for viewing and filtering Additional Codes
+    """
+
+    queryset = (
+        models.ImportBatch.objects.all()
+        .order_by("-created_at")
+        .annotate(
+            chunks_done=Count(
+                "chunks", filter=Q(chunks__status=models.ImporterChunkStatus.DONE)
+            ),
+            chunks_running=Count(
+                "chunks", filter=Q(chunks__status=models.ImporterChunkStatus.RUNNING)
+            ),
+            chunks_waiting=Count(
+                "chunks", filter=Q(chunks__status=models.ImporterChunkStatus.WAITING)
+            ),
+            chunks_errored=Count(
+                "chunks", filter=Q(chunks__status=models.ImporterChunkStatus.ERRORED)
+            ),
+        )
+    )
+    template_name = "importer/list.jinja"
+    filterset_class = ImportBatchFilter
+
+
+class UploadTaricFileView(RequiresSuperuserMixin, FormView):
+    form_class = forms.UploadTaricForm
+    fields = ["name", "split_job"]
+    success_url = reverse_lazy("import_batch-ui-list")
+    template_name = "importer/create.jinja"
+
+    def form_valid(self, form):
+        form.save(user=self.request.user)
+        return super().form_valid(form)
