+++ quotas/models.py
@@ -1,6 +1,5 @@
 from decimal import Decimal
 
-from django.contrib.postgres.fields import JSONField
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
